safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A5_T1-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Cannot use the undeclared label 'LABEL_ANOTHER_LOOP'.
S12.8_A5_T1

PASS Expected parsing failure

TEST COMPLETE
